move get_segments in utils
