+++ espnet2/gan_tts/utils/__init__.py
@@ -1,0 +1,2 @@
+from espnet2.gan_tts.utils.get_random_segments import get_random_segments  # NOQA
+from espnet2.gan_tts.utils.get_random_segments import get_segments  # NOQA
